requirements.txt
Flask==3.1.2
gunicorn
python-dotenv==1.2.1
supabase==2.27.0
groq==1.0.0
langchain-community==0.4.1
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
fastembed
pypdf==6.5.0
httpx==0.28.1
langsmith